requirements.txt
simplejson==3.3.2
requests==2.2.0
Pillow==2.3.0
